Run Jacoco instrumentation out-of-process for Bazel tests
When running Jacoco instrumentation with classloader isolation
we were hitting quite a lot of MissingResourceException when
running tests from Bazel. This is most likely a Gradle issue
(https://github.com/gradle/gradle/issues/9221), and they are
working on a fix. In the meantime, they advised that running
Jacoco out-of-process should fix the issue.

This CL does not change how we run Jacoco in production, as
we received no reports that resemble b/130596259.

BUG: 130596259
Test: existing jacoco tests
Change-Id: I4476aa482d4aa465952a3ed366b1c301e36269a4