Whitelists guava's listenablefuture not use constraints for test.
com.google.guava:listenablefuture:1.0 was extract from Guava but it's still
in the guava artifact as well. To avoid the same class on the classpath
guava depends on version
  com.google.guava:listenablefuture:9999.0-empty-to-avoid-conflict-with-guava
which is empty, so that version promotion removes the duplicated class
from the classpath.

However since AGP 3.3 we have strict version matching between app and
test.
The following scenario breaks:
- app depends on some AndroidX libraries that depend on listenablefuture
 (which was extracted for them)
- test depends on guava somehow.

Now the constraints cannot be resolved.

The fix is to whitelist the artifact to not setup constraints for
it for the test/prod sync.

Change-Id: I1b06a142274c58bf52703bcd52bd9db2f0652496
Fixes: 131431257
Test: new GuavaWhitelistTest