Use Bazel/Foundry to Process Coverage Data (but don't switch it on yet)
The main goal of this is to use Foundty to parallelize Jacoco processing
of Bazel target level data in order to fix the coverage aggregation bug.

Using Jacoco to combine data from multiple targets is prone to errors
for unknown reasons, but combing LCOV data formatted data works fine.
However, doing the Jacoco processing on a target level granularity is
pretty slow due to all the classfile processing and JVM startups, but we
can move the process into Bazel and leverage Foundry to speed it up.

This change involves:
1. Make coverage tests visible to the @results workspace for deploy jars
  - includes passing visibility options through kotlin_test
2. Automatically construct build rules to produce LCOV tracefiles from
whatever outputs zips are present
  - includes inlining the BUILD file for @results since it becomes a
  single function call
  - includes a version of the Jacoco XMl -> LCOV tracefile converter
  that uses stdin and stdout (new version to avoid breaking the existing
  system before we switch)
3. Define a coverage_report rule and an example report to cover the
shard / extra exec matrix
  - includes a LCOV tracefile merger script that can run in Foundry
4. Create a helper shell script for building a defined coverage report

Test: N/A
Bug: 127528212
Bug: 122854984
Change-Id: I7180cb09858611b81e855b9cc04f0403820f1b23