Do not generate R.jar if android.generateRJava=true
With this flag we should revert back to previous behavior, where R.java was
generated and compiled together with the rest of the sources. With this
CL it is possible to offer temporary workaround for e.g. b/130965921,
where third-party plugin was not able to handle project classes coming
from jars.

BUG: 134147459
Test: NonNamespacedApplicationLightRClassesTest
Change-Id: I5fe20428fa92cb9694f0141adb5170c8db791fcb