Fix Json generation dependency check bug
In this bug, the target name is changed inside a CMakeLists.txt that
is transitively referenced through add_subdirectory. The fix is to
make all build files a task input.

A fix isn't possible for fork CMake (3.6.0) because we'd have to change
code in that fork.

I also converted CmakeBasicProjectTest to Kotlin.

Bug: 134086362
Test: Unittested
Change-Id: I525b0d4b2b398efe659e9b5f74d6c88d259288a0