Data binding: Optimize resource merging for library modules.
Data binding consumes the layout info output of resource merging.

A. For a variant in a library module, there are actually 2 types of
resource merging tasks:
  1. "Package resources" task: merges the resources from the current
     module.
  2. "Merge resources" task: merges the resources from both the current
     module and its dependencies.

The "merge resources" task should not be needed, but it is currently
kept because external libraries (e.g., the Kotlin Android extensions
plugin) might consume it.

See bug 110412851 for more details.

B. For a variant in a non-library module (or for an Android test
variant), there is currently only 1 "merge resources" task.

Data binding used to depend on the "merge resources" task in all cases.
but it actually only needs the "package resources" task.

This CL lets data binding depend on the "package resources" task for
library modules to improve performance. In the other cases, it still
depends on the "merge resources" task as the "package resources" task is
not created, and creating "package resources" instead of "merge
resources" might break other tasks that need the "merge resources" task.

Bug: 133092984
Bug: 110412851
Test: Existing tests should pass
Change-Id: I81996274a5d0ea6b41d3a764f664b9c5d89e8a52