Support annotations with no aspect in the aspects agent
The aspects agent resolves conflicts between annotations defined in the
rules file that belong to the same group, giving precedence to method-level
annotations.

Sometimes, however, we want to add annotations to a certain group, but
not necessarily define an aspect for it, so the conflict is solved and
no aspect is added for a particular method. A good example of that is
the @AnyThread annotation, which usually does not have an aspect defined
for, but could be used to take precedence in a method belonging to a
class annotated with @Slow or @UiThread.

Test: InstrumentTest and RulesFileTest
Bug: 133300743
Change-Id: Ia63c0d0b5170aa16eca7564d13d31c8c83623f7e