Enable compatibility with maven-publish plugin.
This creates new instance of SoftwareComponent for each variants allowing
publishing different variants as different artifacts.

Because these must be mapped to a Configuration and because they publish
all the artifacts published to the configuration, we have to create new
configurations dedicated to publication that only receive that must
be published.

For libraries, we need an api and a runtime one. Not because
of us publishing different artifacts but because of the difference
in the transitive dependency information.
For App modules, we need one for APK publishing and one for bundle
publishing because they receive different artifacts to publish (and
there is no setting telling the plugin which one is actually used
by the user.)

This updates PublishingSpec to add a new publish() method to
select which item(s) to publish.

For APKs, because it is possible that more than one APK are created,
the artifact that gets published is a zip of the APK folder output
(including the output.json file), plus the mapping file if present.

Fixes: 37055147
Test: New integration tests for apk, bundle, and aar publishings
Change-Id: I6017717508d8bfaf97470cf029ad00fe173cc35f