Move MissingCompileSdkVersionTest and MissingSdkTest to application.
The sdk directory has a special BUILD file to run SdkAutoDownloadTest
which doesn't interact properly with test files in Kotlin. This was
causing the tests inside these two files to never run.

So this moves those two tests to the application suite which support
Kotlin tests without issues.

(We tried to change glob to include .kt files in the sdk BUILD,
it didn't worked.)

Bug: n/a
Test: n/a
Change-Id: I825629f27d6dc9ba46a41d558e7bf76590716222