Ignore non-static lambda visited by the aspects agent
We incorrectly assumed that lambda functions were desugared to static
methods. That's only the case when the lambda does not access any fields
of the object. Otherwise, the desugared function will not be static. For
now, we use the "lambda$" prefix to determine whether a method is
representing a lambda.

Test: InstrumentTest
Bug: N/A
Change-Id: I88c79abf3a47a744819ef7fe966b79bc8ee6121c